Module mylib::listen_sensor

Poll the temperature sensor every 10 seconds. We support 2 types of temperature sensors: (1) BME280 Temperature Sensor, connected to Blue Pill on port SPI1. This sensor is selected if BME280_OFB is defined in syscfg.yml. (2) Blue Pill internal temperature sensor, connected to port ADC1 on channel 16 This sensor is selected if TEMP_STM32 is defined in syscfg.yml. If this is the Collector Node, send the sensor data to the CoAP Server after polling. This is the Rust version of https://github.com/lupyuen/stm32bluepill-mynewt-sensor/blob/rust/apps/my_sensor_app/OLDsrc/listen_sensor.c

Constants

LISTENER_CB
Indicate that this is a listener callback

SENSOR_POLL_TIME
Poll every 10,000 milliseconds (10 seconds)

Functions

get_temperature
Get the temperature value, raw or computed. sensor_data contains the raw or computed temperature. sensor_type indicates whether sensor_data contains raw or computed temperature. We return the raw or computed temperature, as well as the key and value type.

read_temperature
This listener function is called by Mynewt every 10 seconds (for local sensors) or when sensor data is received (for Remote Sensors). Mynewt has fetched the raw or computed temperature value, passed through sensor_data. If this is a Sensor Node, we send the sensor data to the Collector Node. If this is a Collector Node or Standalone Node, we send the sensor data to the CoAP server.
Return 0 if we have processed the sensor data successfully.

start_sensor_listener
For Sensor Node and Standalone Node: Start polling the temperature sensor every 10 seconds in the background. After polling the sensor, call the Listener Function to send the sensor data to the Collector Node (if this is a Sensor Node) or CoAP Server (is this is a Standalone Node). For Collector Node: Start the Listeners for Remote Sensor Otherwise this is a Standalone Node with ESP8266, or a Sensor Node with nRF24L01. Return Ok() if successful, else return Err() with MynewtError error code inside.