Function mylib::mynewt::kernel::os::os_eventq_poll

pub unsafe extern "C" fn os_eventq_poll(
    arg1: *mut *mut os_eventq, 
    arg2: c_int, 
    arg3: u32
) -> *mut os_event

Poll the list of event queues specified by the evq parameter (size nevqs), and return the "first" event available on any of the queues. Event queues are searched in the order that they are passed in the array.

evq: Array of event queues

nevqs: Number of event queues in evq

timo: Timeout, forever if OS_WAIT_FOREVER is passed to poll.

Return: An event, or NULL if no events available