Crate mylib

This crate provides a library for experimenting and testing rust features on a rapid create/test/modify cycle. This crate will provide examples of various complex use cases for traits, implementations, associated types, etc..

This crate is licensed under the MIT license. You can find a copy of the license at the top of src/lib.rs, src/main.rs, in the project root as LICENSE, as well as in the root of the GitHub Repo - rust-playground

This crate's documentation provides some simple examples. Check out Modules for the various types, traits, and other rust features used in this project.

Example enum Error SqlError type.

Example trait SqlSafe type.

Usage

This crate is NOT on crates.io, but can be obtained from my public GitHub Repo - rust-playground. This crate can be used by adding the below lines to your dependencies in your project's Cargo.toml, if you want to have the main.rs and lib.rs in the same project src/ directory.

[lib]
name = "mylib"
path = "src/lib.rs"

[bin]
name = "mybin"
path = "src/main.rs"

and this to your crate root:

extern crate mylib;

After you have cloned a copy of this project from the GitHub repo, you can run any of the following commands from within the project root:

cargo build     # Build the project from src
cargo run       # Build/Run the project
cargo test      # Test the embedded test cases
cargo bench     # Test the performance of the Bench test cases
cargo doc       # Generate this same documentation locally

Example: Validate a string for SQL Injection Safety

To use a public type from this module, add the below to your crate root. This example uses the SqlSafe trait. This trait overloads the &str type with a is_sql_safe() method.

extern crate mylib;

use mylib::traits::SqlSafe;

The example shows how to use the method and handle it's return. You could put this in your fn main() { }

let username = "Robert'); DROP TABLE Students;--?";
//   try valid username
// let username = "Bobby Tables";

match username.is_sql_safe() {
    Ok(name) => println!("Username \"{}\" is safe", name),
    Err(err) => {
        println!("Error: Username \"{}\" is NOT SQL safe!", username);
        println!("Error: {}", err);
    },
};

assert!("username;".is_sql_safe().is_err());
assert!("username".is_sql_safe().is_ok());

The result will look something like this in the console output after cargo run:

Error: Username "Robert'); DROP TABLE Students;--?" is NOT SQL safe!
Error: found unsafe sql characters in sql string `');'
Error: sql syntax not allowed

Modules

errors
Error Types exported by this library

server
This is documentation for the mylib::server module.

traits
Traits exported by this library

types
Types exported by this library